pyproject.toml
[project]
name = "django-template"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=5.2.1",
    "django-filter>=25.1",
    "django-redis>=5.4.0",
    "djangorestframework>=3.16.0",
    "drf-spectacular>=0.28.0",
    "stripe>=12.1.0",
]
